sjm: port to native sha256
Instead of using lib/hmac.py and lib/hashlib and sha512 in native
python use a simpler HMAC_256 implementation that uses the sha256
native code from micropython.

With that the `sjm.test()` code takes only ~250ms instead of the
more than ~3000ms that the native sha512 needs on an esp32 system.