install,open: make cli/open.py work as a micropython app too
This can be installed via:
```
$ ./install.sh ./cli/open.py
```
and works on e.g. a m5stack echo that uses pin 39 for the pin.
It's configured to activate when pressed for 1s or more.

Note that interrupts on the m5stack are not reliable, see
e.g. https://forum.micropython.org/viewtopic.php?f=2&t=9239